requirements.txt
psycopg2==2.8.6
psycopg2-binary==2.8.6
django_rest_framework